Add target coordinate system parameter values
The stereo_image_proc module generates a 3D point cloud from
stereo images. This patch adds an enum that defines
different types of coordinate systems.
These definitions are made in order to be able
to configure in which coordinate system the 3D point cloud
is represented. The definitions can bu used in nodes and nodelts
such as nodeles/stereo_image_proc.cpp
or nodelets/point_cloud2.cpp.

The conversion methods are needed:
toInt/fromInt: Needed to be able to pass parameters from the
main ros node (stereo_image_proc.cpp) to the nodelet (point_cloud2.cpp)
since passing enum parameters is not possible.
toString/fromString: Needed to be able to parse the textual parameters
passed to the main ros node.

Change-Id: I5dfadfd5692031a86b710761ac4674f1a26d610c