Generate PointCloud2 for the given target coordinate system
Before this patch, the point cloud was implicitly always generated
in the EAST_UP_SOUTH coordinate system. In this patch,
the point cloud is generated in the coordinate system
specified by the target_coordinate_system_ member.

To this end, a switch statement on
the member target_coordinate_system_ is inserted.
The existing code for generating the point cloud
in the EAST_UP_SOUTH coordinate system is copied into
the EAST_UP_SOUTH part of the switch statement.
The EAST_NORTH_UP part of the switch statement
contains corresponding new code.

Change-Id: If8492c9a72103679737dccf25696defb0bbe1889