Target coordinate system is configurable by parameter
Before this commit, the member target_coordinate_system_ in
PointCloud2Nodelet was always set to the default value.

After this commit, in the stereo_image_proc node,
we read a node parameter called target_coordinate_system.
The value is passed to the attached PointCloud2Nodelet
via the shared_params variable. We use the parameter in the
PointCloud2Nodelet to set the member target_coordinate_system_.

Change-Id: Iea7c167030d40923b7f5c177d7aa9115fdae328c