Parallelized filling of color and unprojected points using OpenMP
Before this patch, the 3D point data and the color data
of points_msg were filled sequentially.

This patch uses OpenMP to parallelize filling of 3D and
point cloud data. These two operations write to
separate parts of the memory block reserved for points_msg
and can therefore be parallelized.

Change-Id: Ie51fb055a6d9fa2816cfb7b74ca28958f9776f09