Unproject and fill points into the PointCloud2Ptr in one pass
We reduce the amount of data copies needed to
unproject the disparity image to a 3D point cloud.

Before this commit, a temporary 3D point cloud was
generated from the disparity image using the function
projectDisparityImageTo3D(..). The resulting temporary point
cloud was then copied from the member points_mat_
to the variable mat. To the best of my knowledge,
this copy is unneeded and irrelevant. After the copy,
the points were copied again to the target point cloud.

After this commit, the points are unprojected and stored
in the target point cloud in a single pass without
doing any extra data copies.

Change-Id: I180ce35ceab488e51a095ba735408af5924292c4